requirements.txt
django==4.2
djangorestframework==3.14
pillow==10.0
jdatetime==4.1
pytest==7.4
pytest-django==4.5 